requirements.txt
# This is a list of pypi packages to be installed into virtualenv. Alternatively,
# you can install these as RPMs instead of pypi packages. See the dependecies
# with:
# $ rpmspec -q --requires libcheckb.spec
# $ rpmspec -q --buildrequires libcheckb.spec
#
# A note for maintainers: Please keep this list in sync with libcheckb.spec.
# This list is in the same order as the requirements in the specfile.

# libcheckb
ansible >= 2.3
# bodhi-client is currently not installable from pip: https://github.com/fedora-infra/bodhi/issues/2401
# bodhi-client >= 3.10
configparser >= 3.5.0b2  # needed just for Python 2
Jinja2 >= 2.8
munch >= 2.0.2
progressbar2 >= 3.39
resultsdb_api >= 2.0.0
requests >= 2.7.0
xunitparser >= 1.3.3
PyYAML >= 3.11

# Test suite requirements
mock >= 2.0.0
pytest >= 2.7.3
pytest-cov >= 2.2.1
tox >= 3.13.2

# Documentation requirements
Sphinx >= 1.2.3
sphinx_rtd_theme >= 0.1.9
